make a reusable component to handle sliding animation
